Actualizar archivo Monedero.php para cambiar el nombre de la función "usuario" a "user"

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\UserController;
 use App\Http\Controllers\MonederoController;
 use App\Http\Controllers\ComercioController;
 use App\Http\Controllers\TransaccionController;
+use App\Models\Transaccion;
 use Illuminate\Support\Facades\Route;
 use Whoops\Run;
 
@@ -20,7 +21,7 @@ Route::get('/', function () {
 
 //Rutas de la aplicacion
 // Route::view('/inicio', 'inicio')->name('inicio');
-Route::view('/admin', 'admin')->name('admin');
+Route::view('/admin', 'admin.index')->name('admin')->middleware('auth')->middleware('can:admin');
 
 
 //Usuarios
@@ -31,7 +32,7 @@ Route::view('/admin', 'admin')->name('admin');
 // Route::patch('/users/{user}', [UserController::class, 'update'])->name('user.update');
 // route::delete('/users/{user}', [UserController::class, 'destroy'])->name('user.destroy');
 // Route::post('/users', [UserController::class, 'store'])->name('user.store');
-Route::resource('users', UserController::class)
+Route::resource('/admin/users', UserController::class)
     ->names('users')
     // ->parameters(['users' => 'user'])
     // ->except('create', 'store', 'show')
@@ -39,20 +40,24 @@ Route::resource('users', UserController::class)
     ->middleware('can:manage-users');
 
 //Monederos
-Route::get('/monederos', [MonederoController::class, 'index'])->name('monederos.index');
-Route::get('/monederos/create', [MonederoController::class, 'create'])->name('monedero.create');
-Route::get('/monederos/{monedero}', [MonederoController::class, 'show'])->name('monedero.show');
-Route::get('/monederos/{monedero}/edit', [MonederoController::class, 'edit'])->name('monedero.edit');
-Route::patch('/monederos/{monedero}', [MonederoController::class, 'update'])->name('monedero.update');
-route::delete('/monederos/{monedero}', [MonederoController::class, 'destroy'])->name('monedero.destroy');
+Route::get('/admin/monederos', [MonederoController::class, 'index'])->name('monederos.index')->middleware('auth')->middleware('can:manage-monederos');
+Route::get('/admin/monederos/create', [MonederoController::class, 'create'])->name('monederos.create')->middleware('auth')->middleware('can:manage-monederos');
+Route::get('/admin/monederos/{monedero}', [MonederoController::class, 'show'])->name('monederos.show')->middleware('auth')->middleware('can:manage-monederos');
+Route::get('/admin/monederos/{monedero}/edit', [MonederoController::class, 'edit'])->name('monederos.edit')->middleware('auth')->middleware('can:manage-monederos');
+Route::patch('/admin/monederos/{monedero}', [MonederoController::class, 'update'])->name('monederos.update')->middleware('auth')->middleware('can:manage-monederos');
+route::delete('/admin/monederos/{monedero}', [MonederoController::class, 'destroy'])->name('monederos.destroy')->middleware('auth')->middleware('can:manage-monederos');
+
+Route::get('/user/monedero/{user}', [MonederoController::class, 'monederoUser'])->name('monedero.user')->middleware('auth');
+Route::get('/user/comercios/{user}', [ComercioController::class, 'comerciosUser'])->name('comercio.user')->middleware('auth');
+Route::get('/user/transaccion/create', [TransaccionController::class, 'create'])->name('transaccion.create')->middleware('auth')->middleware('can:crear-transacciones');
 
 //Transacciones
-Route::get('/transacciones', [TransaccionController::class, 'index'])->name('transacciones.index');
-Route::get('/transacciones/create', [TransaccionController::class, 'create'])->name('transacciones.create');
-Route::get('/transacciones/{transaccion}', [TransaccionController::class, 'show'])->name('transacciones.show');
-Route::get('/transacciones/{transaccion}/edit', [TransaccionController::class, 'edit'])->name('transacciones.edit');
-Route::patch('/transacciones/{transaccion}', [TransaccionController::class, 'update'])->name('transacciones.update');
-Route::delete('/transacciones/{transaccion}', [TransaccionController::class, 'destroy'])->name('transacciones.destroy');
+Route::get('/admin/transacciones', [TransaccionController::class, 'index'])->name('transacciones.index')->middleware('auth')->middleware('can:manage-transacciones');
+Route::get('/admin/transacciones/create', [TransaccionController::class, 'create'])->name('transacciones.create')->middleware('auth')->middleware('can:manage-transacciones');
+Route::get('/admin/transacciones/{transaccion}', [TransaccionController::class, 'show'])->name('transacciones.show')->middleware('auth')->middleware('can:manage-transacciones');
+Route::get('/admin/transacciones/{transaccion}/edit', [TransaccionController::class, 'edit'])->name('transacciones.edit')->middleware('auth')->middleware('can:manage-transacciones');
+Route::patch('/admin/transacciones/{transaccion}', [TransaccionController::class, 'update'])->name('transacciones.update')->middleware('auth')->middleware('can:manage-transacciones');
+Route::delete('/admin/transacciones/{transaccion}', [TransaccionController::class, 'destroy'])->name('transacciones.destroy')->middleware('auth')->middleware('can:manage-transacciones');
 
 
 //Comercios
@@ -65,17 +70,17 @@ Route::delete('/transacciones/{transaccion}', [TransaccionController::class, 'de
 //ADMINISTRADOR
 // Route::get('/admin', [AdminController::class, 'index'])->name('admin.index');
 
-Route::resource('comercios', ComercioController::class)
+Route::resource('/admin/comercios', ComercioController::class)
     ->names('comercios')
     ->middleware('auth')
     ->middleware('can:manage-comercios');
 
-Route::resource('sorteos', ComercioController::class)
+Route::resource('/admin/sorteos', ComercioController::class)
     ->names('sorteos')
     ->middleware('auth')
     ->middleware('can:manage-sorteos');
 
-Route::resource('promociones', ComercioController::class)
+Route::resource('/admin/promociones', ComercioController::class)
     ->names('promociones')
     ->middleware('auth')
     ->middleware('can:manage-promociones');
@@ -86,6 +91,8 @@ Route::get('/dashboard', function () {
     return view('dashboard');
 })->middleware(['auth', 'verified'])->name('dashboard');
 
+
+//Rutas de perfil
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
